refactor(view): Reorganize directories for budget input view Move BudgetInputView into subdirectories
