+++ src/main/java/lotto/view/input/BudgetInputVIew.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.view.input;
 
 import camp.nextstep.edu.missionutils.Console;
 import lotto.utils.validator.BudgetValidator;
